Rework the application to be configured using what Play calls 'compile-time dependency injection' which is akin to Spring Java Config.
Doing so has exposed the org.OrgContext companion object for what it is: a holder of configuration and singleton instances that gets passed around between anyone and everyone. One by one, each property of the OrgContext class should be removed until every component is only injected with exactly the values or collaborators it requires.

Metrics are now available on /non-public/monitoring/metrics. When deploying this application you should disallow users to that url using nginx, apache etc.